python/requirements.txt
PyJWT==2.10.1
PyJWT==2.10.1
Requests==2.32.3
elementpath